% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_onehot.R
\name{catto_dummy}
\alias{catto_dummy}
\title{Dummy encoding}
\usage{
catto_dummy(train, ..., test)
}
\arguments{
\item{train}{The training data, in a \code{data.frame} or \code{tibble}.}

\item{...}{The columns to be encoded.  If none are specified, then
all character and factor columns are encoded.}

\item{test}{The test data, in a \code{data.frame} or \code{tibble}.}
}
\value{
The encoded dataset in a \code{matrix} if all character and
  factor columns have been encoded, otherwise the encoded dataset in
  a \code{data.frame} or \code{tibble}, whichever was input.  If a test
  dataset was provided, a named list is returned holding the encoded
  training and test datasets.
}
\description{
Dummy encoding
}
\examples{
catto_dummy(iris, response = Sepal.Length)
}
