% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_freq.R
\name{catto_mean}
\alias{catto_mean}
\title{Mean encoding}
\usage{
catto_mean(train, ..., response, test, verbose = TRUE)
}
\arguments{
\item{train}{The training data, in a \code{data.frame} or \code{tibble}.}

\item{...}{The columns to be encoded.  If none are specified, then
all character and factor columns are encoded.}

\item{response}{The response variable used to calculate means.}

\item{test}{The test data, in a \code{data.frame} or \code{tibble}.}

\item{verbose}{Should informative messages be printed?  Defaults to
\code{TRUE}.}
}
\value{
The encoded dataset in a \code{data.frame} or \code{tibble},
  whichever was input.  If a test dataset was provided, a named list
  is returned holding the encoded training and test datasets.
}
\description{
Mean encoding
}
\examples{
catto_mean(iris, response = Sepal.Length)
}
