% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cattonum.R
\docType{package}
\name{cattonum}
\alias{cattonum}
\title{cattonum: Encode Categorical Features}
\description{
Functions for dummy encoding, frequency encoding,
label encoding, leave-one-out encoding, mean encoding,
median encoding, and one-hot encoding.
}
