% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{cattonum_df2}
\alias{cattonum_df2}
\title{Constructor for class cattonum_df2}
\usage{
cattonum_df2(train = NULL, test = NULL)
}
\arguments{
\item{train}{\code{NULL} (the default), or a tibble or data.frame.}

\item{test}{\code{NULL} (the default), or a tibble or data.frame
with the same names as \code{train}.}
}
\value{
A list of class \code{cattonum_df2} with names "train"
and "test".
}
\description{
Constructor for class cattonum_df2
}
\examples{
cattonum_df2()
}
