% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy-onehot.R
\name{catto_dummy}
\alias{catto_dummy}
\title{Dummy encoding}
\usage{
catto_dummy(train, ..., test, verbose = TRUE)
}
\arguments{
\item{train}{The training data, in a \code{data.frame} or \code{tibble}.}

\item{...}{The columns to be encoded.  If none are specified, then
all character and factor columns are encoded.}

\item{test}{The test data, in a \code{data.frame} or \code{tibble}.}

\item{verbose}{Should informative messages be printed?  Defaults to
\code{TRUE} (not yet used).}
}
\value{
The encoded dataset in a \code{cattonum_df} if no test dataset was
provided, and the encoded datasets in a \code{cattonum_df2} otherwise.
}
\description{
Dummy encoding
}
\examples{
catto_dummy(iris)
}
