% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{catto_loo}
\alias{catto_loo}
\title{Leave-one-out encoding}
\usage{
catto_loo(train, ..., response = NULL, test = NULL, verbose = TRUE)
}
\arguments{
\item{train}{The training data, in a \code{data.frame} or \code{tibble}.}

\item{...}{The columns to be encoded.  If none are specified, then
all character and factor columns are encoded.}

\item{response}{The response variable used to calculate means.}

\item{test}{The test data, in a \code{data.frame} or \code{tibble}.}

\item{verbose}{Should informative messages be printed?  Defaults to
\code{TRUE}.}
}
\value{
The encoded dataset in a \code{cattonum_df} if no test dataset was
provided, and the encoded datasets in a \code{cattonum_df2} otherwise.
}
\description{
Leave-one-out encoding
}
\examples{
catto_loo(iris, response = Sepal.Length)
}
